Fix rare, intermittent, endless loop in timeouts
Don't reduce millis() and timeout_start_ms to 16 bit as this causes
wrapping every 65 seconds which can lead to endless loops from
checkTimeoutExpired() if the sensor times out when the 16 bit millis() is
within io_timeout of wrapping.

+++ VL53L0X.cpp
@@ -16,7 +16,7 @@
 #define startTimeout() (timeout_start_ms = millis())
 
 // Check if timeout is enabled (set to nonzero value) and has expired
-#define checkTimeoutExpired() (io_timeout > 0 && ((uint16_t)millis() - timeout_start_ms) > io_timeout)
+#define checkTimeoutExpired() (io_timeout > 0 && (millis() - timeout_start_ms) > io_timeout)
 
 // Decode VCSEL (vertical cavity surface emitting laser) pulse period in PCLKs
 // from register value
